Update README.md
SeekConnect is a platform designed to make reporting lost items or missing persons easy and private. It keeps everything organized, so there's no confusion.
"When someone finds a lost item or has information about a missing person, they can share it on SeekConnect for everyone to see. This helps reunite lost items with their owners and find missing people faster."